% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datachile_package.R
\name{datachile-exports}
\alias{datachile-exports}
\alias{tibble}
\alias{as_tibble}
\alias{\%>\%}
\alias{select}
\alias{mutate}
\alias{distinct}
\alias{unnest_}
\alias{map}
\alias{fromJSON}
\alias{user_agent}
\alias{modify_url}
\alias{GET}
\alias{http_type}
\alias{content}
\alias{http_error}
\alias{status_code}
\title{datachile exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the datachile package to avoid listing dependences of datachile
}
