% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datachile_package.R
\docType{package}
\name{datachile}
\alias{datachile}
\alias{datachile-package}
\title{Tools to simplify getting tidy data from DataChile API.}
\description{
Tools to simplify getting tidy data from DataChile API.
}
\author{
Mauricio Vargas
}
