% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get_cube_info.R
\name{get_cube_info}
\alias{get_cube_info}
\title{Obtain cube's details}
\usage{
get_cube_info(path)
}
\arguments{
\item{path}{Must be a valid cube name that can be accesed by running `get_cubes()`}
}
\description{
Obtain cube's details
}
\keyword{functions}
