% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_get_regiones.R
\name{get_geo_units}
\alias{get_geo_units}
\title{Obtain geographical units (comunas and regions)}
\usage{
get_geo_units()
}
\description{
Obtain geographical units (comunas and regions)
}
\keyword{functions}
